Rely on System.currentTimeMillis only (not nanoTime) due to JENKINS-19778
On most Jenkins servers, System.nanoTime gives a more accurate reading of
elapsed time, because it is unaffected when the clock is changed. However,
on the Jenkins server in issue 19778, System.nanoTime jumps forwards
several times during the build, by about 30-40 seconds each time. Once the
build has finished, the Timestamper plugin reports an elapsed time of about
twice what it should be.

It does not seem possible for the plugin to reliably determine whether it
can use System.nanoTime, especially when running within a virtual machine.

Overall, it is not worthwhile for the plugin to rely on System.nanoTime.