[FIXED JENKINS-32074] When rootURL is empty, store cookies at '/'
When Jenkins is running from the top-level URL, the rootURL variable is an empty
string. In this case, store the cookies at '/' instead of rootURL.

Prior to this change, the cookies were created with an empty path, and the
browser defaults to storing them at the build URL. These cookies will now be
deleted because they would override the cookies stored at the Jenkins URL.

Bug introduced by JENKINS-29899.