[FIXED JENKINS-30143] Store elapsed time in timestamp console notes
The timestamp console notes previously relied on the context being a Run, so
that they can get the build start time and use it to calculate the elapsed time.
The elapsed time is now stored within the console notes during the build.

This allows timestamps to be displayed when a different context type is
provided, e.g. the workflow plugin provides a StepAtomNode.

Builds that ran with an earlier version of Timestamper and an unrecognised
context type will now display the system clock timestamps but not the elapsed
timestamps, because that information was not recorded.