Refresh Javascript file when the plugin is upgraded/downgraded
ConsoleAnnotatorFactory serves script.js with a 1 day expiry. If the Timestamper
plugin was upgraded/downgraded in that time, the out-of-date script would still
be executed by the browser.

Work around this by moving the code into annotator.js, and having script.js
download annotator.js from a different URL each time the plugin version changes.
The annotator.js file is served from /static and so has a 1 year expiry.