[JENKINS-38085] Revert to default way of loading the script and CSS
A custom approach was previously used to load the script and CSS for displaying
the timestamps. It caused the script and CSS to be reloaded when the Timestamper
plugin is upgraded or downgraded, rather than being cached for the next day.

It seems this doesn't work for everyone (JENKINS-38085), so I'm reverting back
to the default behaviour. Note that it is known to work with HTTPS and through a
HTTPS proxy (JENKINS-35315) at least for some users.

As a result, the TimestampAnnotatorFactory class will have to be manually
renamed each time the script or CSS is changed in a new release.