Revert "[JENKINS-72690] Make the widget sticky (#287)"
When other panels have been added to the side bar, they are shown
overlapping with the stick timestamper panel which has a transparent
background.  That is noticeable with the Collapsible Section plugin
which can generate a fairly long panel.

I understand the use case for making it sticky, but the overlapping text
is concerning.

This reverts commit b2e06b50de137bc74a5016cf49f27686b45a814e.